Temp: 상품상세 알레르기 유발 요인 추가
