+++ src/components/content/product/productNutritionInfo.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import classNames from 'classnames/bind';
-import styles from '../../styles/pages/productPage.module.scss';
+import styles from '../../../../styles/pages/productPage.module.scss';
 
 const cx = classNames.bind(styles);
